math.pi has only first 15 digits valid
The math.pi contains only some approximation of pi. Actually only the first 15 digits are valid.
I think you should trim to first 15 digits only (next one is 2 so rounding is correct) or to preserve output more consistent with production output the valid 42 digits.
For the IPv4 format, you should probably round up to .27 instead of .26 because next digits are 53....
IPv6 is fine.