fetch-ifsc: Ensure output directory exists before fetching IFSC data.
Previously, running the `scripts/fetch-ifsc.sh data/ifsc` command
as suggested in `config.sample.toml` failed with an error:

tar: could not chdir to 'data/ifsc'

as the target directory did not exist.

This commit adds a `mkdir -p` for the destination path to ensure
the download and extraction succeed.